+++ codeForDoc/src/quick_start/queryAndOnSnapshot.ts
@@ -9,38 +9,40 @@ import {
 	limit,
 } from 'firelordjs'
 
-export const dummy = async () => {
-	await getDocs(
-		query(
-			example.collection(),
-			where('f.h', '>', 1010 as const),
-			orderBy('f.h'),
-			limit(10)
-		)
+getDocs(
+	query(
+		example.collection(),
+		where('f.h', '>', 1010 as const),
+		orderBy('f.h'),
+		limit(10)
 	)
+).then(querySnapshot => {
+	querySnapshot.docChanges().forEach(docChange => {
+		console.log(docChange.doc)
+		console.log(docChange.type)
+		console.log(docChange.oldIndex)
+		console.log(docChange.newIndex)
+	})
+	querySnapshot.forEach(docSnapshot => {})
 
-	const unsub = onSnapshot(
-		query(
-			example.collectionGroup(),
-			where('b.d', 'array-contains', { e: 'hello' }),
-			orderBy('f.g'),
-			startAfter(new Date())
-		),
-		querySnapshot => {
-			querySnapshot.forEach(docSnapshot => {
-				const data = docSnapshot.data()
-			})
-		},
-		error => {},
-		{ includeMetadataChanges: false }
-	)
+	querySnapshot.docs.forEach(docSnapshot => {})
+})
 
-	const unsub2 = onSnapshot(
-		example.doc('abc'),
-		documentSnapshot => {
-			const data = documentSnapshot.data()
-		},
-		error => {},
-		{ includeMetadataChanges: true }
-	)
-}
+const unsub = onSnapshot(
+	query(
+		example.collectionGroup(),
+		where('b.d', 'array-contains', { e: 'hello' }),
+		orderBy('f.g'),
+		startAfter(new Date())
+	),
+	querySnapshot => {},
+	error => {},
+	{ includeMetadataChanges: false }
+)
+
+const unsub2 = onSnapshot(
+	example.doc('abc'),
+	docSnapshot => {},
+	error => {},
+	{ includeMetadataChanges: true }
+)