+++ src/utilForTests.ts
@@ -174,7 +174,7 @@ export const readThenCompareWithWriteData = async (
 	await expect(getDocFromCache(ref)).rejects.toThrow()
 }
 
-export const writeThenReadTest = async (
+export const writeThenCompareWithRead = async (
 	writeCallback: (
 		data: ReturnType<typeof generateRandomData>
 	) => Promise<DocumentReference<User>>